Move over RewriterStepSequencer to frontend module
Also moves its implementations
