+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/ast/conditions/normalizedEqualsArguments.scala
@@ -19,7 +19,7 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_2.ast.conditions
 
-import org.neo4j.cypher.internal.compiler.v3_2.tracing.rewriters.Condition
+import org.neo4j.cypher.internal.frontend.v3_2.helpers.rewriting.Condition
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{Equals, Expression, FunctionInvocation, Property, functions}
 
 case object normalizedEqualsArguments extends Condition {
