+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/ast/conditions/orderByOnlyOnVariables.scala
@@ -20,7 +20,7 @@
 package org.neo4j.cypher.internal.compiler.v3_2.ast.conditions
 
 import org.neo4j.cypher.internal.frontend.v3_2.ast.{Variable, OrderBy, SortItem}
-import org.neo4j.cypher.internal.compiler.v3_2.tracing.rewriters.Condition
+import org.neo4j.cypher.internal.frontend.v3_2.helpers.rewriting.Condition
 
 case object orderByOnlyOnVariables extends Condition {
   def apply(that: Any): Seq[String] = {
