+++ community/cypher/frontend-3.2/src/main/scala/org/neo4j/cypher/internal/frontend/v3_2/helpers/rewriting/RewriterTaskBuilder.scala
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.internal.compiler.v3_2.tracing.rewriters
+package org.neo4j.cypher.internal.frontend.v3_2.helpers.rewriting
 
 import org.neo4j.cypher.internal.frontend.v3_2.Rewriter
