+++ community/cypher/frontend-3.2/src/test/scala/org/neo4j/cypher/internal/frontend/v3_2/helpers/rewriting/RewriterStepSequencerTest.scala
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.internal.compiler.v3_2.tracing.rewriters
+package org.neo4j.cypher.internal.frontend.v3_2.helpers.rewriting
 
 import org.neo4j.cypher.internal.frontend.v3_2.Rewriter
 import org.neo4j.cypher.internal.frontend.v3_2.test_helpers.CypherFunSuite
